Move measurement APIs to new directory
