+++ hume/__init__.py
@@ -1,8 +1,16 @@
 """Module init."""
+
 from importlib.metadata import version
 
-from hume._batch import BatchJob, BatchJobDetails, BatchJobState, BatchJobStatus, HumeBatchClient, TranscriptionConfig
-from hume._stream import HumeStreamClient, StreamSocket
+from hume._measurement.batch import (
+    BatchJob,
+    BatchJobDetails,
+    BatchJobState,
+    BatchJobStatus,
+    HumeBatchClient,
+    TranscriptionConfig,
+)
+from hume._measurement.stream import HumeStreamClient, StreamSocket
 from hume.error.hume_client_exception import HumeClientException
 
 __version__ = version("hume")
