+++ hume/_measurement/batch/batch_job.py
@@ -1,15 +1,16 @@
 """Batch job."""
+
 import json
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Union
 
-from hume._batch.batch_job_details import BatchJobDetails
-from hume._batch.batch_job_status import BatchJobStatus
 from hume._common.retry_utils import RetryIterError, retry
+from hume._measurement.batch.batch_job_details import BatchJobDetails
+from hume._measurement.batch.batch_job_status import BatchJobStatus
 from hume.error.hume_client_exception import HumeClientException
 
 if TYPE_CHECKING:
-    from hume._batch.hume_batch_client import HumeBatchClient
+    from hume._measurement.batch.hume_batch_client import HumeBatchClient
 
 
 class BatchJob:
