+++ hume/_measurement/batch/batch_job_details.py
@@ -1,11 +1,12 @@
 """Batch job details."""
+
 import json
 from datetime import datetime
 from typing import Any, Dict, List, Optional
 
-from hume._batch.batch_job_state import BatchJobState
-from hume._batch.batch_job_status import BatchJobStatus
 from hume._common.config_utils import config_from_model_type
+from hume._measurement.batch.batch_job_state import BatchJobState
+from hume._measurement.batch.batch_job_status import BatchJobStatus
 from hume.error.hume_client_exception import HumeClientException
 from hume.models import ModelType
 from hume.models.config.model_config_base import ModelConfigBase
