+++ hume/_measurement/stream/__init__.py
@@ -1,0 +1,9 @@
+"""Module init."""
+
+from hume._measurement.stream.hume_stream_client import HumeStreamClient
+from hume._measurement.stream.stream_socket import StreamSocket
+
+__all__ = [
+    "HumeStreamClient",
+    "StreamSocket",
+]
