--- hume/_stream/__init__.py
@@ -1,8 +1,0 @@
-"""Module init."""
-from hume._stream.hume_stream_client import HumeStreamClient
-from hume._stream.stream_socket import StreamSocket
-
-__all__ = [
-    "HumeStreamClient",
-    "StreamSocket",
-]
